+++ R/GSMVARconstruction.R
@@ -35,11 +35,11 @@
 #'          \emph{Journal of Econometrics}, \strong{192}, 485-498.
 #'    \item Kalliovirta L. and Saikkonen P. 2010. Reliable Residuals for Multivariate Nonlinear
 #'          Time Series Models. \emph{Unpublished Revision of HECER Discussion Paper No. 247}.
-#'    \item Virolainen S. (forthcoming). A statistically identified structural vector autoregression with endogenously
-#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}.
 #'    \item Virolainen S. 2022. Gaussian and Student's t mixture vector autoregressive model with application to the
 #'      asymmetric effects of monetary policy shocks in the Euro area. Unpublished working
 #'      paper, available as arXiv:2109.13648.
+#'    \item Virolainen S. 2025. A statistically identified structural vector autoregression with endogenously
+#'          switching volatility regime. \emph{Journal of Business & Economic Statistics}, \strong{43}, 1, 44-54.
 #'  }
 #' @examples
 #' # GMVAR(1, 2), d=2 model:
@@ -225,8 +225,8 @@ GSMVAR <- function(data, p, M, d, params, conditional=TRUE, model=c("GMVAR", "St
 #'  \itemize{
 #'    \item Kalliovirta L., Meitz M. and Saikkonen P. 2016. Gaussian mixture vector autoregression.
 #'          \emph{Journal of Econometrics}, \strong{192}, 485-498.
-#'    \item Virolainen S. (forthcoming). A statistically identified structural vector autoregression with endogenously
-#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}.
+#'    \item Virolainen S. 2025. A statistically identified structural vector autoregression with endogenously
+#'          switching volatility regime. \emph{Journal of Business & Economic Statistics}, \strong{43}, 1, 44-54.
 #'    \item Virolainen S. 2022. Gaussian and Student's t mixture vector autoregressive model with application to the
 #'      asymmetric effects of monetary policy shocks in the Euro area. Unpublished working
 #'      paper, available as arXiv:2109.13648.
@@ -426,8 +426,8 @@ alt_gsmvar <- function(gsmvar, which_round=1, which_largest, calc_cond_moments=T
 #'    \item Muirhead R.J. 1982. Aspects of Multivariate Statistical Theory, \emph{Wiley}.
 #'    \item Kalliovirta L., Meitz M. and Saikkonen P. 2016. Gaussian mixture vector autoregression.
 #'          \emph{Journal of Econometrics}, \strong{192}, 485-498.
-#'    \item Virolainen S. (forthcoming). A statistically identified structural vector autoregression with endogenously
-#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}.
+#'    \item Virolainen S. 2025. A statistically identified structural vector autoregression with endogenously
+#'          switching volatility regime. \emph{Journal of Business & Economic Statistics}, \strong{43}, 1, 44-54.
 #'    \item Virolainen S. 2022. Gaussian and Student's t mixture vector autoregressive model with application to the
 #'      asymmetric effects of monetary policy shocks in the Euro area. Unpublished working
 #'      paper, available as arXiv:2109.13648.
@@ -560,11 +560,11 @@ gsmvar_to_sgsmvar <- function(gsmvar, calc_std_errors=TRUE, cholesky=FALSE) {
 #'    \item Muirhead R.J. 1982. Aspects of Multivariate Statistical Theory, \emph{Wiley}.
 #'    \item Kalliovirta L., Meitz M. and Saikkonen P. 2016. Gaussian mixture vector autoregression.
 #'          \emph{Journal of Econometrics}, \strong{192}, 485-498.
-#'    \item Virolainen S. (forthcoming). A statistically identified structural vector autoregression with endogenously
-#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}.
 #'    \item Virolainen S. 2022. Gaussian and Student's t mixture vector autoregressive model with application to the
 #'      asymmetric effects of monetary policy shocks in the Euro area. Unpublished working
 #'      paper, available as arXiv:2109.13648.
+#'    \item Virolainen S. 2025. A statistically identified structural vector autoregression with endogenously
+#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}, \strong{43}, 1, 44-54.
 #'  }
 #' @examples
 #' \donttest{