+++ R/loglikelihood.R
@@ -175,8 +175,8 @@
 #'          \emph{Springer}.
 #'    \item McElroy T. 2017. Computation of vector ARMA autocovariances.
 #'          \emph{Statistics and Probability Letters}, \strong{124}, 92-96.
-#'    \item Virolainen S. (forthcoming). A statistically identified structural vector autoregression with endogenously
-#'           switching volatility regime. \emph{Journal of Business & Economic Statistics}.
+#'    \item Virolainen S. 2025. A statistically identified structural vector autoregression with endogenously
+#'          switching volatility regime. \emph{Journal of Business & Economic Statistics}, \strong{43}, 1, 44-54.
 #'    \item Virolainen S. 2022. Gaussian and Student's t mixture vector autoregressive model with application to the
 #'      asymmetric effects of monetary policy shocks in the Euro area. Unpublished working
 #'      paper, available as arXiv:2109.13648.