+++ R/simulate.R
@@ -7,7 +7,7 @@
 #' @param nsim number of observations to be simulated.
 #' @param seed set seed for the random number generator?
 #' @param ... currently not in use.
-#' @param init_values a size \eqn{(pxd)} matrix specifying the initial values, where d is the number
+#' @param init_values a size \eqn{(p\times d)} matrix specifying the initial values, where d is the number
 #'   of time series in the system. The \strong{last} row will be used as initial values for the first lag,
 #'   the second last row for second lag etc. If not specified, initial values will be drawn according to
 #'   mixture distribution specifed by the argument \code{init_regimes}.
@@ -24,20 +24,18 @@
 #' @param ntimes how many sets of simulations should be performed?
 #' @param drop if \code{TRUE} (default) then the components of the returned list are coerced to lower dimension if \code{ntimes==1}, i.e.,
 #'   \code{$sample} and \code{$mixing_weights} will be matrices, and \code{$component} will be vector.
-#' @param girf_pars This argument is used internally in the estimation of generalized impulse response functions (see \code{?GIRF}).
-#'   You should ignore it (specifying something else than null to it will change how the function behaves).
 #' @details The argument \code{ntimes} is intended for forecasting: a GMVAR, StMVAR, or G-StMVAR process can be forecasted by simulating
 #'  its possible future values. One can easily perform a large number simulations and calculate the sample quantiles from the simulated
 #'  values to obtain prediction intervals (see the forecasting example).
 #' @return If \code{drop==TRUE} and \code{ntimes==1} (default): \code{$sample}, \code{$component}, and \code{$mixing_weights} are matrices.
 #'   Otherwise, returns a list with...
 #'   \describe{
-#'     \item{\code{$sample}}{a size (\code{nsim}\eqn{ x d x }\code{ntimes}) array containing the samples: the dimension \code{[t, , ]} is
+#'     \item{\code{$sample}}{a size (\code{nsim}\eqn{\times d \times}\code{ntimes}) array containing the samples: the dimension \code{[t, , ]} is
 #'      the time index, the dimension \code{[, d, ]} indicates the marginal time series, and the dimension \code{[, , i]} indicates
 #'      the i:th set of simulations.}
-#'     \item{\code{$component}}{a size (\code{nsim}\eqn{ x }\code{ntimes}) matrix containing the information from which mixture component
+#'     \item{\code{$component}}{a size (\code{nsim}\eqn{\times}\code{ntimes}) matrix containing the information from which mixture component
 #'      each value was generated from.}
-#'     \item{\code{$mixing_weights}}{a size (\code{nsim}\eqn{ x M x }\code{ntimes}) array containing the mixing weights corresponding to
+#'     \item{\code{$mixing_weights}}{a size (\code{nsim}\eqn{\times M \times}\code{ntimes}) array containing the mixing weights corresponding to
 #'      the sample: the dimension \code{[t, , ]} is the time index, the dimension \code{[, m, ]} indicates the regime, and the dimension
 #'      \code{[, , i]} indicates the i:th set of simulations.}
 #'   }
@@ -84,17 +82,43 @@
 #'        probs=c(0.025, 0.5, 0.972))
 #' @export
 
-simulate.gsmvar <- function(object, nsim=1, seed=NULL, ..., init_values=NULL, init_regimes=1:sum(gsmvar$model$M),
-                            ntimes=1, drop=TRUE, girf_pars=NULL) {
+simulate.gsmvar <- function(object, nsim=1, seed=NULL, ..., init_values=NULL, init_regimes=1:sum(object$model$M),
+                            ntimes=1, drop=TRUE) {
+  simulate_gsmvar_int(object=object, nsim=nsim, seed=seed, ..., init_values=init_values, init_regimes=init_regimes,
+                      ntimes=ntimes, drop=drop)
+}
+
+
+#' @title INTERNAL Simulate method for class 'gsmvar' objects
+#'
+#' @description \code{simulate_gsmvar_int} an internal a simulation function for class 'gsmvar' objects.
+#'   It allows to simulate observations from a GMVAR, StMVAR, or G-StMVAR process.
+#'
+#' @inheritParams simulate.gsmvar
+#' @param girf_pars This argument is used internally in the estimation of generalized impulse response functions (see \code{?GIRF}).
+#'   You should ignore it (specifying something else than null to it will change how the function behaves). Should be a list with the following elements:
+#'   \describe{
+#'     \item{\code{shock_numb}}{an integer indicating which shock to impose in the GIRF estimation.}
+#'     \item{\code{shock_size}}{a numeric value indicating the size of the structural shock to impose in the GIRF estimation.}
+#'     \item{\code{include_mixweights}}{a logical value indicating whether the GIRF should be estimated for the mixing weights as well.}
+#'   }
+#' @inherit simulate.gsmvar references return details seealso
+#' @keywords internal
+
+simulate_gsmvar_int <- function(object, nsim=1, seed=NULL, ..., init_values=NULL, init_regimes=1:sum(gsmvar$model$M),
+                                ntimes=1, drop=TRUE, girf_pars=NULL) {
+  ## girf_pars for GIRFs use:
   # girf_pars$shock_numb - which shock?
   # girf_pars$shock_size - size of the structural shock?
   # girf_pars$include_mixweights - should GIRFs be estimated for the mixing weights as well? TRUE or FALSE
+
   # If !is.null(girf_pars) and girf_pars$include_mixweights == TRUE, returns a size (N+1 x d+M) vector containing
   #                                                                  the estimated GIRFs for the variables and
   #                                                                  and the mixing weights (column d+m for the m:th regime).
   # If !is.null(girf_pars) and girf_pars$include_mixweights == FALSE, returns a size (N+1 x d) vector containing
   #                                                                   the estimated GIRFs for the variables only.
   # The first row for response at impact
+
   gsmvar <- object
   if(!is.null(seed)) set.seed(seed)
   epsilon <- round(log(.Machine$double.xmin) + 10)