+++ NEWS.md
@@ -285,3 +285,4 @@
 
 * GFEVDs are now calculated as the average over the GFEVDs based on the different initial values (previously calculated based on the average
   of the GIRFs based on the different initial values).
+* Updated the references.